gcp-tracing/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace
